+++ types/int.go
@@ -17,6 +17,10 @@ func NewInt(x int64) Int {
 	return Int{*new(big.Int).SetInt64(x)}
 }
 
+func (x Int) Add(y Int) (Int) {
+	return Int{*x.Int.Add(&x.Int, &y.Int)}
+}
+
 func (x Int) GetBSON() (interface{}, error) {
 	return x.Int.String(), nil
 }